pyproject.toml
[project]
name = 'python-getpaid-bitpay'
version = '0.1.0'
description = 'BitPay payment gateway integration for python-getpaid ecosystem.'
readme = 'README.md'
license = {text = 'MIT'}
authors = [
    {name = 'Dominik Kozaczko', email = 'dominik@kozaczko.info'},
]
requires-python = '>=3.12'
classifiers = [
    'Development Status :: 3 - Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
    'Topic :: Office/Business :: Financial',
    'Topic :: Office/Business :: Financial :: Point-Of-Sale',
    'Typing :: Typed',
]
dependencies = [
    'python-getpaid-core>=0.1.0',
    'httpx>=0.27.0',
    'ecdsa>=0.19.0',
    'anyio>=4.0',
]

[dependency-groups]
dev = [
    'pytest>=8.0',
    'pytest-asyncio>=0.24.0',
    'pytest-cov>=5.0',
    'respx>=0.22.0',
    'ruff>=0.9.0',
    'pre-commit>=4.0',
    'ty>=0.0.16',
]
docs = [
    'furo>=2024.8.6',
    'sphinx>=8.0',
    'myst-parser>=4.0',
]

[project.urls]
Homepage = 'https://github.com/django-getpaid/python-getpaid-bitpay'
Repository = 'https://github.com/django-getpaid/python-getpaid-bitpay'
Documentation = 'https://getpaid-bitpay.readthedocs.io/'
Changelog = 'https://github.com/django-getpaid/python-getpaid-bitpay/releases'

[project.entry-points."getpaid.backends"]
bitpay = 'getpaid_bitpay.processor:BitPayProcessor'

[build-system]
requires = ['hatchling']
build-backend = 'hatchling.build'

[tool.hatch.build.targets.wheel]
packages = ['src/getpaid_bitpay']

[tool.pytest.ini_options]
testpaths = ['tests']
asyncio_mode = 'auto'

[tool.coverage.run]
branch = true
source = ['getpaid_bitpay']

[tool.coverage.report]
show_missing = true

[tool.ruff]
target-version = 'py312'
line-length = 80
src = ['src', 'tests']

[tool.ruff.lint]
select = [
    'E',    # pycodestyle errors
    'W',    # pycodestyle warnings
    'F',    # pyflakes
    'I',    # isort
    'N',    # pep8-naming
    'UP',   # pyupgrade
    'B',    # flake8-bugbear
    'A',    # flake8-builtins
    'SIM',  # flake8-simplify
    'TC',   # type-checking imports
    'RUF',  # ruff-specific
]
ignore = [
    'N818',  # Exception names match getpaid-core convention
    'B027',  # Empty methods in ABC are intentional default no-ops
]

[tool.ruff.lint.per-file-ignores]
'tests/**' = ['TC001', 'RUF012']

[tool.ruff.lint.isort]
force-single-line = true
lines-after-imports = 2
known-first-party = ['getpaid_bitpay']

# --- ty type checker ---
[tool.ty.environment]
python-version = '3.12'

[tool.ty.terminal]
error-on-warning = true

[[tool.ty.overrides]]
include = ['tests/**']
[tool.ty.overrides.rules]
unresolved-attribute = 'ignore'
invalid-argument-type = 'ignore'
